Project-expense-tracking/requirements.txt
streamlit==1.49.1
fastapi==0.116.2
pydantic==2.11.7
uvicorn==0.35.0
mysql-connector-python==9.4.0
pandas==1.5.3
requests==2.32.4
pytest==8.4.2